category finish

+++ internal/service/categories.go
@@ -1,6 +1,9 @@
 package service
 
-import "forum/internal/models"
+import (
+	"fmt"
+	"forum/internal/models"
+)
 
 func (s *service) GetCategories() ([]*models.Category, error) {
 	categories, err := s.repo.GetCategories()
@@ -30,9 +33,15 @@ func (s *service) PostCategoryPost(post_id int, categoryIds []int) error {
 }
 
 func (s *service) CreateCategory(form models.CategoryCreateForm) error {
+	if form.Name == "" {
+		return fmt.Errorf("Blank category name")
+	}
 	return s.repo.CreateCategory(form.Name)
 }
 
 func (s *service) DeleteCategory(id int) error {
+	if id <= 0 {
+		return fmt.Errorf("invalid category ID")
+	}
 	return s.repo.DeleteCategory(id)
 }
